drop "returns?" information from source code's Expr tree
Differentiate between the "everything is an expression" view of
the language, and the "no unreachable code" view expected by the
bytecode verifier.  Now a `throw` can appear anywhere where a
regular expression is permitted, and is treated like a void
expression.

This makes parsing and bytecode emit slightly simpler, at the
cost of an `as-returning-code` rewrite step that maps from the
source text to the code attribute representation.

Expr changes: Empty expression is again respresented by a
singleton Expr instead of a singleton constant value.  This is
slightly less expressive and leads to less concise bytecode in
some cases.  Except for no-arg instructions, there is now a
`with-argv` function to change return type or arguments.  There
is a new predicate `strict-evaluation?`.

Parsing changes: `jb/no-return` no longer exists and is subsumed
by `jb/void`, e.g. as return type of throw or recur.  The
wildcard handler has been removed from the representation of
TryCatchFinally.

Bytecode changes: Specialized handling of constant branch has
been removed.