insert two copies of finally statements only
If the try/catch/finally has a non-empty finally statement, then
regular control flow of both the try body and all catch handlers exits
through the end of the respective block.  The blocks contain no
xRETURN instruction (by construction), have no non-local `recur`
(forbidden by the language), and all jumps and branches target labels
within the block.

This means that a goto at the end of each block can jump to a copy of
the finally statement, placed right after the "catch any" handler that
holds the other copy.  With this there are always two copies of
finally, independent of the number of exception handlers.

javac uses a different strategy: it embeds copies of the finally
statement before any exit from the try body or a catch handler.  Let's
see how the non-javac approch pans out...