eliminate dependency on BlockCodeBuilderImpl.isEmpty()
The current setup guarantees that the try block's region is never
empty (it holds at least a xRETURN or a GOTO), which implies that the
try-catch region is also never empty.  This means that the two
isEmpty() checks can be avoided, but with the "cost" that an empty try
block is not omitted entirely.

This makes code generation slightly simpler by not catering for a
rather obscure edge case.  It removes the need to use isEmpty(), which
is right now from the "impl" package, and --add-open on impl can be
dropped.  With isEmpty() gone, there is also no need to work with a
instance of BlockCodeBuilder and the indirection via block() can be
omitted.