emit namespace segment after every macro definition
Once the segment is written, the macro's value is immediately
delivered to its promise.  This reduces the time spent waiting by the
compilation thread.

For the core library, this drives up the segment count significantly.
For most namespaces, it does not make a difference.

Going back to the old behavior of holding macros back until used is an
option.  But then care must be taken, because the non-deterministic
mode may have enqueued additional macro definitions that cannot be
flushed immediately.