add flag collect-macro-definitions? for deterministic mode
Reduces the number of namespace segments when building the AoT class
files for tinyclj.core.

The emit thread deadlocks when combining this with non-deterministic
compilation: msg-emit-tlf processing would issue a message to the emit
thread and then wait for the result.  Which is the situation that the
just removed prepare thread and its fast track message bypassing
prevented -- at the cost of reordered messages.

So now there is the slow mode distinguished by deterministic and
compact output.  And then there is the default mode, which is
non-deterministic and geared toward compilation throughput.