change interface functions to anonymous only
Method code wrapped via LambdaMetafactory has no access to "this", the
reference to its own instance.  Additionally, there is no way to patch
such an instance with references to letfn* functions appearing later
in the source code.  Both of this limitations can be made tangible if
instance functions are forbidden to have an inner name that is visible
to its arity body.

Special form syntax following in the footsteps of `:once`:

  (^ItfClass fn* [...] ...)

Possible sugared syntax for such anonymous function literals:

  ^ItfClass (fn [...] ...)
  (defn ^ItfClass foo [...] ...)

Using type hint syntax reads slightly better than the low-level fn*
syntax's prefixed type tag.  Emacs also copes better with this
variant.

The defn would be slightly problematic in that the name `foo` is only
used for the def, but not for the implicit fn form.  People will
stumble over `foo` not being visible for the defn's arity.