reject narrowing numeric type conversions again
The rewrite of type adaptions introduced code that allow a narrowing
of e.g. a double value to float type.  Reject this again, and also
conversion of char to int.

When emitting code, compensate for the special case char -> int, which
is a no-op on the bytecode level.  Char values are subsumed under
"computational int" types, and the rules of the source language do not
apply here.