fix endless recursion of tclj1's known-good? test when using bootstrap-tcljc
When the core lib's requires are recorded as is, then there is a
dependency loop "tinyclj.core" -> "tinyclj.core.protocols" ->
"tinyclj.core".

Because compilation always starts with core lib and its requires are
subsumed into core, make sure that the core lib never has any recorded
dependencies.  But add the source files of the erased requires to
those of core.