drop `defn` support to define an interface function
While the syntax `(defn ^Itf foo [...] ...)` is convenient, it is too
misleading compared to regular function definition.  It would define
an entity `foo` that is not a function, and the name `foo` cannot be
referenced within the method body.

So drop this shortcut (reversing part of commit 304c4f81e4) and switch
to longhand syntax instead.  The bilk of this commit is the mechanical
conversion to def+fn syntax.