compile time evaluation of `str` on any constant value
Move constant descriptor construction into the bytecode emitter, and
change the frontend to store the actual value in Expr's `const-value`
member.  The only exception is ConstantDescs/NULL, which remains the
replacement for the value nil.

Quote processing is still a bit verbose, because it needs to unpack
any embedded syntax quotes.  For the non-quote case, collection
literals construct a constant collection when possible.

Compile time evaluation of `str` becomes easier, with the caveat that
it does not expand a LazySeq instance on its own.