generate unique name for clashing fn name from form's string
Hashing the form directly does not work if the form contains an
instance of java.util.regex.Pattern.  In general, two instances of
Pattern are not equivalent even if they are created from the same
pattern string.  This means that e.g. (.hashCode #"foo") will return a
different value each time it is compiled.

The workaround to hash the string of the form instead is more
expensive.  Maybe find a cheaper solution to generate stable names?
Alternatively the reader could produce a symbolic description instead
of a Pattern instance, but this would deviate from Clojure's
behaviour.