redefine macro assert with ~(pr-str x) after (load "core_print")
The base definition of assert is too early, and must push the call to
pr-str to runtime.

Assertions after the second definition take fewer bytecodes to fill in
the AssertionError instance by loading a single string constant (~2
bytes).  The constant pool holds the string with the duplicated
"Assert failed: " prefix.

Compare this to the first definition: load mh for pr-str, construct
list instance, call mh with it, map result to "" if it's nil, and
finally concatenate with prefix (~17 bytes).  The constant pool holds
the construction recipe for the list.