refactor: replace reader implementation
The new implementation should be easier to follow, and has a very
small interface to the parser.  The hierarchical state tracking in
start/end is a little bit tricky, but this is compensated by the
simpler code around it.

Performance seems to be similar to the prior version.  Line count of
reader and parser has decreased by around 15%.