fix type conflict in recur of IChunkedSeq's internal-reduce function
The first argument `s` is the "receiver" of internal-reduce and has
the type IChunkedSeq.  The recur passes the result of (chunk-next s)
to `s`, but this can be any ISeq, like PersistentList.

Fixed by wrapping the function body in a `loop` and setting type ISeq
for the loop variable `s`.