work around method resolve confusion around MethodTypeDesc/of
A recent update added a new overload for the method.  Now the
bootstrap compiler (which I don't want to change anymore) thinks the
call (MethodHandleDesc/of <ClassDesc> <ClassDesc>) can be satisfied by
either one of

  static MethodTypeDesc of(ClassDesc returnDesc, ClassDesc... paramDescs)

or

  static MethodTypeDesc of(ClassDesc returnDesc, List<ClassDesc> paramDescs)

The first would wrap the argument in an array, the second one would
cast the interface argument from ClassDesc to List.  The latter
happens because Clojure adds the required cast automatically, unlike
Java where it must be provided.

Instead of changing the bootstrap compiler's behavior, add a wrapper
function for the intended method to resolve the ambiguity.

Also clean up `rt/mtd-of` and prefer this function over manually
putting together the MTD.