refactor: move interfacing to OS/JVM into helper classes
The old code passed a lot of state along to enable actions further
down the call chain.  These actions are often done for their
side-effects, like reading a file, defining a class, or storing class
bytes for unit tests.

Try to identify the actions, add them as interface methods to the
relevant IO abstraction, and encapsulate the values on which the
actions depend.

This removes state from the old carrier classes, and better describes
which actions are taken on the namespace level, during general code
compilation, and by the top-level build setup.

Performance is either unchanged or slightly better.