shout if function fails to submit a task
There is a small chance that an ExecutorService is shut down and then
a late function thread tries to submit a nested function it encounters
during regular compilation.  Make sure that a stack trace is printed
in this case.

Possible fixes:

Use StructuredTaskScope as intended.  The scope would be the top-level
form.  Any functions encountered while compiling the form it would
fork from this scope.  Calling join on the scope would wait for tasks
initiated after the call to join as well.

Use an atomic counter that is incremented before a function task is
started, and decremented once it completes.  Only shut down the
ExecutorService if the counter is zero.  It may be possible to use a
reverse j.u.c.Semaphore with MAX_VALUE permits for this.  Then an
acquire/release around a task would never block, and shutdown would be
able to do an acquire(MAX_VALUE) first.