fix deftype-form* timeout if nested function has error
`deftype-form*` waits for the defined class before the compilation
thread can continue.  Make sure that the emitter thread always
delivers a response to this request, even if compilation of a local
function throws an error and the class cannot be emitted.

The fix is to make `complete-req!` more robust by wrapping it in a
try/catch that sends back the exception instance if message processing
throws.