remove StringBuffer from list of default imports
Quoting StringBuilder's API documentation: "This class provides an API
compatible with StringBuffer, but with no guarantee of synchronization.
This class is designed for use as a drop-in replacement for StringBuffer
in places where the string buffer was being used by a single thread (as is
generally the case).  Where possible, it is recommended that this class be
used in preference to StringBuffer as it will be faster under most
implementations."

Given that StringBuffer is rarely used, retract its shortcut.  Now it must
be pulled in by a dedicated import entry or referenced via its longhand
name java.lang.StringBuffer.