add support for Java records
If a `deftype` lists java.lang.Record in its specs, then the deftype's
class is emitted as a Java record.  Some of the compile time checks
given in the Java Language Specification are implemented:

* no field name must be on the list of prohibited names

* no field must be declared mutable

* the :abstract option must not be set

If a user provided no-arg method matches a field name, or a method
tries to define one of the Object methods equals() or hashCode(), then
the verifier will reject the class file.

The method toString() can be defined as part of the deftype form, and
defaults to the generic implementation provided by
java.lang.runtime.ObjectMethods/bootstrap otherwise.