use -XX:TieredStopAtLevel=1 for one-shot compilation
Compiler bootstrap is not running long enough for C2 compilation to
pay off.  Even more so when JVM ergonomics for 8+8+0 cores sets
CICompilerCount=10 (instead of 8+0+0's 4).  CPU event samples of
async-profiler under CompileBroker::compiler_thread_loop() when doing
200 back to back bootstraps: 23160 (55.47%) with 8+0+0, a whopping
54435 (69.40%) with 8+8+0, and 1895 (5.41%) with TieredStopAtLevel=1.
For the single bootstrap of stageFI1 this gives a speedup of 1.6 to
0.9s, and for stageDI1 a speedup of 1.3 to 1.4s.

One-shot compilation of regular application code seems to benefit as
well.  I am seeing a speedup of 1.5 for a spot check (from 0.75s to
0.50s).  The different setup doesn't seem to change much in this
regard: The core namespace is AOT compiled, which is a large piece of
rather sequential work in the bootstrap scenario.  The core macros
come as part of the application class loader (and as such would
benefit from Project Leyden as well).  On the other hand: with targets
watch-and-compile or watch-and-test being the usual mode of operation,
it is probably not worth the effort to tune applications' Makefiles
for the "run once" scenario.