pick up Clojure changes to Delay and LazySeq in clojure-1.12.0-alpha5
Upstream Clojure commit is 5dbf142e0d

Delay.java is copied over as is.  The patch to LazySeq replaces the
IFn member with an instance of Supplier (field, constructor, and value
realization in method force()).

This replaces uses of synchronized with an ReentractLock.  This means
that forcing a Delay or the value of a LazySeq no longer pins the
current virtual thread.

The upstream Clojure patch takes the bootstrap time of stageFI[12]
from 0.90s to 0.86s -- a speedup of around 1.05.

Also update the license terms to match the upstream commit 6975553804,
and include the missing terms for the use of Guava Murmur3 hash.