force namespace segment clinit via Lookup.ensureInitialized()
This drops an empty synthetic method from namespace classes and
replaces the implicit "initialize class by touching something static"
approach with an explicit call to Lookup.ensureInitialized().  This
method was introduced as part of JDK 15.

There is considerable churn in the tests to remove the forceClinit~1
methods.  The few calls to forceClinit~1 are replaced with
tinyclj.lang.RT.ensureInitialized().

One slight complication of this change: req-flush-namespace-segment
must produce a class even if it is empty, or the "force clinit"
expression it reports back points nowhere.  This is done by appending
an empty top-level form prior to passing the segment to write-segment.