experiment: automatic unboxing when assigning to primitive type
Support assignment of compatible reference type to primitive type.
The checkcast required is simple to derive, as is the method to
extract the requested primitive value.  No widening or narrowing is
performed in the process.  That is, if the target type is `int`, then
the value *must* be of type `Integer`.  Any other type of value is
rejected, if possible at compile time but surely at runtime.

An interesting consequence of this change is that primitive type hints
can be used with destructuring.  A form like

  (let [[^int x] [1 2 3]]
    nil))

gets expanded to something like this

  (let* [vec__1549 [1 2]
         ^int x (clojure.core/nth vec__1549 0 nil)]
    nil)

which assigns an Object to `^int x`, and the regular assignment logic
takes care of the rest.

This is a small part of the intended changes when Valhalla lands.  I'm
jumping the gun here, but this is basic Java semantics and the
compiler changes are not too intrusive.  It saves cognitive effort for
the simplest situations involving boxed to unboxed conversion.