refactor: replace member `form` of Expr with `line-number`
This means that the form that produced an Expr instance is no longer
available in general.  Only the form's line number, or rather its
usual approximation, can be retrieved.

For BindingRef introduce a member `simple-nm` that is equivalent to
the former contents of `form`, but typed as c.l.Symbol.

This change has some impact on error messages, because the form of the
source expression cannot be produced as part of the displayed message.
This is not a problem if the destination of an operation is a
BindingRef with a `simple-nm`.  Lets see how this pans out...

Next step: make Expr's line-number immutable.