refactor: move most primitive function implementations to predefined.cljt
Only primitive functions that are tightly coupled to compile-form
remain in compiler.cljt.

Clean up the implementations in the process, introducing a central
facility for type conversion with namespace tcljc.compiler.adapt and
extending tcljc.type accordingly.  Chosen names are a bit on the
verbose side, but may need more tuning in the future still.

Drop truthyReference() because it only reimplements booleanCast().