turn tcljc.type.Wrapper into a record
The bootstrap compiler tclj-in-tclj turns this into a class extending
java.lang.Record *without* a RecordAttribute, but the JVM doesn't seem
to have a problem with this kind of non-record.

tcljc itself produces a proper record with RecordAttribute.  Which
means the fields of Wrapper can be trusted by the JVM and can
contribute to optimizations.  This being a compiler I doubt that the
functions in question become hot enough to make a difference.  It
still may lessen the impact of replacing magic constants, so do that.