rewrite method/constructor invoke to handle varargs directly
Prior code delegated to MethodHandle.invoke() to match the argument
list to the invoked method's paramter list.

Drop this special case to create the required varargs array, store the
arguments in the array, and use the array as last argument of the
method invocation.

To stay compabtible with the old setup, primitives passed e.g. to
String/format must be converted to their boxed representation.

Using bootstrap-check-with-tcljc as a spot check, performance seems be
back to the old baseline with this change.  No idea if this is caused
by the revamped method resolve strategy, or the changed bytecode
emitted by tcljc.  This is with minimal compilation and a short
running program, so performance indications are not meaningful anyway.