do not match up nil argument with primitive parameter
That the compiler would allow such a constellation because of its
generous auto unboxing is rather distracting when resolving Java
methods.  A misaligned argument list can produce a rather strange
error message in this case.  But keep unboxing nil in general
assignment situations for now!

The JVM does not surface null_type.  It exists in the StackMapTable
attribute of methods, and jshell/javac seem to use the concept as
well.  As a type it shares some properties with Object (compatible
with any reference) and other properties with Void (only valid value
is null).  Lacking such a type, argument/parameter matchup cannot rely
solely on Class instances to decide valid pairing.  (In former
compiler incarnations I tried to introduce a dedicated class for
null_type, but this required special treatment all over the code
base.)

In namespace invoke, introduce ArgSig to reify an argument signature
covering both argument types and the arguments' nil-ness.  This leads
to a bit of mechanical code churn, because the atypes array is
replaced with an instance of ArgSig.