establish error context around compilation of expanded form
In general, macro expansion does not propagate line numbers into
nested forms.  This means that in a code snippet like

  (defn f []
    (for [i [1 2 3]]
      xxx))

the error when resolving `xxx` gets no line number from the deeply
nested expansion of the `for` macro.

As a fix, establish a dedicated error context when compiling the
result of an expand-macro call.  If the form prior to expansion comes
from the reader, then it does have a line number.  This line will be
applied to the error exception's meta data.