support casting of primitive value to reference type
Similar to the "reference to primitive" direction, only an instance of
the primitive's box type can be produced this way.  The result's
static type is that of the type hint.  That is, if `foo` is of type
int, then the static type of "^Object foo" is Object and not Integer.