elide constructor invocation as fn arity
With the recent addition of Class/new there is shorthand syntax to
create a function instance wrapping a constructor MethodHandle.  So
drop special casing of arity bodies invoking a constructor and elide
these methods as well.

As a consequence, `deftype`'s positional factory function `->T` is no
longer materialized as a method.  This changes for example the code
emitted by deftype-test/data-only-factory-and-alias.  An invocation of
the factory function is now translated to a NEW-DUP-INVOKESPECIAL
instead of the slightly shorter INVOKESTATIC.

If this turns out to be undesireable, the `new` form constructed by
`build-positional-factory` in file core_deftype.cljt can be wrapped as
(identity (new ...)).  This would emit a dedicated method body like
before.