only elide method definition if arity and handle agree on varargs
The `varargs?` flag of an imported static function's arity is
reconstructed from the isVarargsCollector() of its MethodHandle.  This
means that the arity's method definition can only be replaced with an
existing method if both agree with respect to their isVarArgs()
status.

The error scenario goes like this: a vararg function arity is replaced
by mistake with a MethodHandle that takes a plain array instead.  This
function is imported into another namespace, where its signature is
reconstructed as taking a non-vararg array as last argument.  Invoking
the function then fails because the arguments no longer match up.

A Method or Constructor's varargs information is not part of its
symbolic DirectMethodHandleDesc representation.  So add an additional
member `varargs?` to InvokeXref to pass it along to the consuming
function tcljc.emitter.prepare/arity-code-static.

Also fix publics.cljt to never set the mutually exclusive `varargs?`
and `variadic?` members at the same time.  For an importing namespace,
a variadic function arity looks just like an `Object...` varargs one.